pyproject.toml
[project]
name = "openstack-images-sync"
authors = [
    { name = "Openstack Charmers", email = "openstack-charmers@lists.launchpad.net" },
]
version = "0.1.0"
description = "Sync images from a simplestreams source to an openstack cloud"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache 2" }
classifiers = [
    "Framework :: FastAPI",
    "Environment :: Web Environment",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operation System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "pydantic-settings",
    "python-simplestreams",
    "python-keystoneclient",
    "python-glanceclient",
    "click",
]

[project.optional-dependencies]
dev = ["ruff"]

[project.scripts]
openstack-images-sync = "openstack_images_sync.__main__:cli"

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py310"
line-length = 99

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "B008", # do not perform function calls in argument defaults
]
